% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{vsts_auth_key}
\alias{vsts_auth_key}
\title{Visual Studio Team Service Authentication Key}
\usage{
vsts_auth_key(user, pass)
}
\arguments{
\item{user}{username to access VSTS project}

\item{pass}{password to access VSTS project}
}
\description{
Visual Studio Team Service Authentication Key
}
