% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{vsts_get_releases}
\alias{vsts_get_releases}
\title{Visual Studio Project Release Information}
\usage{
vsts_get_releases(domain, project, auth_key, query = NULL)
}
\arguments{
\item{domain}{the location of the visual studio server}

\item{project}{the name of the project in \code{domain} to look at}

\item{auth_key}{authentication key generated by using \code{\link{vsts_auth_key}}}

\item{query}{a list of extra parameters that can be sent to the API call:
\describe{
\item{\code{definitionId}}{[integer] Releases from this release definition Id.}
}}
}
\description{
These functions will allow you to scrape release information from Visual Studio.
}
