% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{vsts_get_projects}
\alias{vsts_get_projects}
\title{Azure DevOps Projects}
\usage{
vsts_get_projects(domain, auth_key, quiet = FALSE)
}
\arguments{
\item{domain}{The name of the Azure DevOps organization.}

\item{auth_key}{Authentication key generated by using \code{\link{vsts_auth_key}}}

\item{quiet}{logical whether want general running information from printing. Any issue with the API call will
still show up if set to \code{TRUE}}
}
\description{
These functions will allow you to scrape project information from Azure DevOps.
}
\details{
For more information about project API calls check
\url{https://docs.microsoft.com/en-us/rest/api/azure/devops/core/Projects}.
}
