% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{vsts_create_release}
\alias{vsts_create_release}
\alias{vsts_get_release}
\title{Visual Studio Project Release Information}
\usage{
vsts_create_release(domain, project, auth_key, body)

vsts_get_release(domain, project, release, auth_key)
}
\arguments{
\item{domain}{the location of the visual studio server}

\item{project}{the name of the project in \code{domain} to look at}

\item{auth_key}{authentication key generated by using \code{\link{vsts_auth_key}}}

\item{body}{a list of extra parameters that can need to be sent to the API call (* mandatory):
\describe{
\item{\code{artifacts *}}{[character] the name of a branch in the repository (cannot combine with \code{commit})}
\item{\code{definitionId *}}{[integer] the id of a commit in the repository (cannot combine with \code{branch})}
\item{\code{description *}}{[character] path of an item in the repository}
\item{\code{isDraft}}{[character] name of the person who committed the change}
\item{\code{manualEnvironments}}{[character] name of the author}
\item{\code{properties}}{[Date] start date to search from}
\item{\code{reason}}{[Date] end date to search from}
}}

\item{release}{ID number of the release to look at}
}
\description{
These functions will allow you to scrape releases information from Visual Studio.
}
