% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{vsts_get_release_defs}
\alias{vsts_get_release_defs}
\title{Visual Studio Project Release Definition Information}
\usage{
vsts_get_release_defs(domain, project, auth_key, quiet = FALSE)
}
\arguments{
\item{domain}{the location of the visual studio server}

\item{project}{the name of the project in \code{domain} to look at}

\item{auth_key}{authentication key generated by using \code{\link{vsts_auth_key}}}

\item{quiet}{logical whether want general running information from printing. Any issue with the API call will
still show up if set to \code{TRUE}}
}
\description{
These functions will allow you to scrape release definition information from Visual Studio.
}
