% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{vsts_get_repos}
\alias{vsts_get_repos}
\alias{vsts_create_repos}
\alias{vsts_delete_repos}
\title{Visual Studio Projects}
\usage{
vsts_get_repos(domain, project, auth_key, quiet = FALSE)

vsts_create_repos(domain, project, repo, auth_key, quiet = FALSE)

vsts_delete_repos(domain, project, repo, auth_key, quiet = FALSE)
}
\description{
Visual Studio Projects
}
