% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\docType{class}
\name{vsts_account}
\alias{vsts_account}
\title{Visual Studio Team Services Account}
\format{An \code{\link{R6Class}} generator object}
\usage{
vsts_account
}
\arguments{
\item{user}{username for the Visual Studio account}

\item{pass}{password for the Visual Studio account}

\item{domain}{domain name for the Visual Studio account}

\item{project}{(optional) project name within the domain of the Visual Studio account}
}
\description{
Visual Studio Team Services Account
}
\examples{
proj <- vsts_account$new('<username>', '<password>', '<domain name>', '<project name>')
str(proj)

}
\keyword{data}
