% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\docType{class}
\name{vsts_account}
\alias{vsts_account}
\title{Visual Studio Team Services Account}
\format{An \code{\link{R6Class}} generator object}
\usage{
vsts_account
}
\description{
Visual Studio Team Services Account
}
\keyword{data}
