% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\docType{data}
\name{vsts_account}
\alias{vsts_account}
\title{Visual Studio Team Services Account}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
vsts_account
}
\description{
In order to call any of the APIs to VSTS, credentials to access the server are needed. Once these have been set-up they won't
need to be used in a session.
}
\keyword{datasets}
