% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{vsts_auth_key}
\alias{vsts_auth_key}
\title{Visual Studio Team Service Authentication Key}
\usage{
vsts_auth_key(user, pass)
}
\arguments{
\item{user}{username to access VSTS project}

\item{pass}{password to access VSTS project}
}
\value{
An authentication key string in the form of 'Basic <Base 64 of user:pass>'
}
\description{
Creation of a VSTS authentication key that will be used when running any of the API calls.
}
\examples{
auth_key <- vsts_auth_key("<username>", "<password>")

}
