% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workitem.R
\name{vsts_create_workitem}
\alias{vsts_create_workitem}
\title{Visual Studio Project Work Items}
\usage{
vsts_create_workitem(domain, project, item_type, auth_key, ...)
}
\arguments{
\item{domain}{the location of the visual studio server}

\item{project}{the name of the project in \code{domain} to look at}

\item{item_type}{the type of work item to be created}

\item{auth_key}{authentication key generated by using \code{\link{vsts_auth_key}}}

\item{...}{arguments passed to \link{vsts_get_workitem_fields}}
}
\description{
These functions will allow you to scrape work item infromation from a particular Visual Studio project.
}
