% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{vsts_get_projects}
\alias{vsts_get_projects}
\title{Visual Studio Projects}
\usage{
vsts_get_projects(domain, auth_key, quiet = FALSE)
}
\arguments{
\item{domain}{the location of the visual studio server}

\item{auth_key}{authentication key generated by using \code{\link{vsts_auth_key}}}

\item{quiet}{logical whether want general running information from printing. Any issue with the API call will
still show up if set to \code{TRUE}}
}
\description{
These functions will allow you to scrape project information from Visual Studio.
}
\details{
For more information about project API calls check
\url{https://docs.microsoft.com/en-us/rest/api/vsts/core/projects}.
}
